server/requirements.txt
Django==6.0
psycopg2==2.9.11
# psycopg2-binary==2.9.10  # for development use
djangorestframework==3.16.0
celery==5.5.3
redis==7.1.0
channels==4.2.2
channels-redis==4.2.1
djangorestframework-simplejwt==5.5.0
pytest==9.0.2
pytest-django==4.11.1
pytest-asyncio==1.1.0
drf-yasg==1.21.10
uvicorn[standard]==0.40.0
gunicorn==23.0.0
pillow==12.0.0
django-redis==6.0.0
daphne==4.2.1